Setup test harness and simple tests of MS_VRH
Make use of the mtest / x_unit unit test framework to
allow testing of functions. To run tests cd into
tests subdir inside matlab and type "runtests"
or "runtests test_MS_VRH" to just test that
function.

The current tests just check that the VRH avarage
of some elastic constants with themselves do not
change the matrix and that an error is thrown if
the arguments do not make sense.

To check errors it is necessary to use the two
argument form of error() in SH_VRH.

To run the tests x_unit must be installed:
http://www.mathworks.com/matlabcentral/fileexchange/22846-matlab-xunit-test-framework
http://www.mathworks.com/matlabcentral/fx_files/22846/1/content/doc/index.html