Minimum change to implement documentation
This puts documentation into the Matlab help
viewer when the msat directory is imported
into the Matlab path. info.xml tells
Matlab to look in the /html directory and
should not need changeing. /html/helptoc.xml
lists the items that go in the table of contents
in the help viewer. text.html is a very minimal
html file so we can see that this has worked.