Simple tool to generate documentation for funcs
Linke the Matlab 'help' command, this extracts
the contents of the first comment block in a
.m file. This is used to build a asciidoc
document (.txt) with the name of the
function (from filename.m) which can be
converted to html (and included in the Matlab
help viewer). An index.html document is also
created with links for all functions.