Optimise MS_phasevels
MS_phasevels is one of the most costly functions in
MSAT. For example, running MS_anisotropy with lma
output and n=10000 (which is what should be used)
takes ~19 seconds. All of this time is spent in
MS_phasevels. Here we reduce the time spent in
some of the minor functions inside MS_phasevels
and reduce ~19s to ~10s. Specifically we:

  * Avoid using the Matlab cross() function for
    two three-vectors and use our own function
    instead (the Matlab function is general and
    thus costly: we know we have two three-vectors).
  * Use two special case vector rotation functions
    instead of a general case function. This avoids
    four matrix multiplies and about 10 trigonometric
    function calls per MS_phasevels evaluation.
  * Remove an argument from cart2: this is never
    used and saves an if statment.

Overall this nearly saves 50% of the execution time. The
profiler now shows that execution time is now dominated
by the velo function (~50% of the time) which is hard to
optimise.