Fix 'reverse' argument for plotting functions
MS_plot and MS_sphere both accept the 'reverse'
argument to (un)reverse the sense of the colourmap.
The default is blue for high and red for low (as
expected for seismic data - i.e. backwards and
reversed as per the Matlab defaults). The reverse
argument now turns off the reversal making red
fast and blue slow. This fixes the bug reported
by Michele Zucali.