Clarify MS_rotEuler and add tests
Make the behaviour of MS_rotEuler be to rotate
the crystal elasticity into the frame of the
sample (the active rotation) not the other
way around. Provide an option to do the
passive rotation. Add tests for this and
to check we get the same Voigt elasticity
as a simple (10 crystal) VPSC calc.