Add tests, unwind output
Make the output consistent with the S&S method. Add tests
following some of those from S&S. Note that in three cases
(tests commented out) the result is a few degrees different.
Shows up more for longer periods. Is this close to a faliure
mode for S&S or is something wrong with the wavelets?