Fix length unit for thickness
MSAT works in units of km/s for velocity so
the thicknesses provided to MS_effective_splitting_N
need to be km. Fix this in the calculation of thickness.
Also, I don't think we need to convert from ray to
geographcal reference frame for the step by step
polarizations. We can go back to the Gaussian
Wavelet as this now gives results in a sensible
time.