Fix calculation of isorropic velocity
Rather than calculating the isotropic velcoty
from the isotropic component of the poly-xtal
anisotropic tensor (the norm of which changes
with texture) use simple VRH mean of the single
xtal tensor. This gives values that are just a
function of the temperature and do not vary with
texture.

Also split the velocity graph into two graphs.