Allow MS_save to write to an already open file
If 'fname' is a real (rather than a string) write
to that assuming it is an fid (and avoid opening a
new file). In this case the file is not closed
prior to exiting MS_save. This can be used when
either several Cijs must be written to the same
file or to create parts of larger reports.

Note that 'fname' = 1 will write data to the
terminal.