Fix bug in MS_axes for hexagonal case
It is possible for hexaginal crystals in the
"proper" orentation to have eigenvalue e(1)
~= e(2) and e(1) ~= e(3) but e(2) == e(3). In
this case the ndistinct function would miss the
final equality and we would treat the symmetry
as orthorhombic (possibly ending up with X(1)
being the distinct direction, if the distinct
direction lined up with the stiff one).

Here we add the missing test such that this
hexagonal case is treated properly. This fixes
github issue #24