Make MS_interpolate work for triclinic
For triclinic crystals we can choose any
axes for interpolation, as long as they
are consistant. Rather than choosing the
bisectrix of the dilational and Voigt
stiffness tensors, we can just use the
dilational tensor. This avoids the
unstable behaviour for a large
triclinic part (c.f. sectoon 3.2 of
Browaeys and Chevrot).

This is implemented by adding an X3_distinct
check in the triclinic case in MS_axes. With
this in place the test for triclinic interpolation
passes, so this is uncommented.

However, we still have problems with interpolation
between crystals of different symmetry. We should
consistently choose to use the dilational stiffness
tensor in all cases, but constructing a reasonable
test case is not totally trivial.