Make sure pdata and sdata arguments are sane
We allow "arrays" of the same length for the
pdata and sdata arguments in MS_plot. We mean
vectors but the way MATLAB handles row and
coloumn vectors is a bit annoying. This
adds an internal function to confirm that
the array is indeed a vector, and to make
the vector a row vector.

We could probably also unroll any 2d arrays
too before carrying on but as yet this is
not done.