Handle arrays as well as vectors for xdata
This works for all the input I can think of,
but it would be nice to formalise tests for
the plotting functions. Anyway, we can now
mix sizes (in the odd MATLAB sense where
a column vector and a row vector each of n
items can be accessed as v(n) but have
different sizes) as long as the number of
elements in all of the input arrays match.
Column vectors are transposed before processing.
Actuall matricies are transformed to vectors
in whatever order v=m(:) does. Help : does not
say exactly what this order is... but I hope it
is always the same.